Create RefusalAdviceQuestionForm
Extract and rename helper methods into a FormBuilder which allows us to
simplify method attributes and the view.
